src/main/java/com/twu/biblioteca/domain/model/Movie.java
package com.twu.biblioteca.domain.model;


import java.util.Date;

public class Movie  {

    private String id;
    private String name;
    private String director;
    private Date year;
    private String rating;

    public Movie(String id, String name, String director, Date year, String rating) {
        this.id = id;
        this.name = name;
        this.director = director;
        this.year = year;
        this.rating = rating;
    }

    public String getId() {
        return id;
    }

    public void setId(String id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getDirector() {
        return director;
    }

    public void setDirector(String director) {
        this.director = director;
    }

    public Date getYear() {
        return year;
    }

    public void setYear(Date year) {
        this.year = year;
    }

    public String getRating() {
        return rating;
    }

    public void setRating(String rating) {
        this.rating = rating;
    }
}
